Add game mode 1 screen and update main menu screen
In this commit, a new XML description for a font and some sprites have been added. MainMenuScreen.cs has been updated with a renamed constant for clarity, and a new file, GameMode1Screen.cs has been created to provide the functionality of a new game mode. The content manager has also been updated to build and manage the new resources.