Add planet implementation and improve game logic
Added new classes for modelling data related to planets' size, points, and textures, and for handling different types of planets in the game. Changes also include improvement in the game mechanics like scoring, visual updates to how the planets appear, and the addition of Verlet point for handling physics. GameMode1Screen class was also updated for setting the content manager for PlanetTextures.